+++ resource/stock_balance/type.go
@@ -12,7 +12,7 @@ type StockBalanceResource interface {
 	FindByBranch(branchID string) ([]model.StockBalance, *error_wrapper.ErrorWrapper)
 	FindByItem(itemID string) ([]model.StockBalance, *error_wrapper.ErrorWrapper)
 	FindByBranchAndItem(branchID, itemID string) (*model.StockBalance, *error_wrapper.ErrorWrapper)
-	Update(branchID, itemID string, currentStock int) *error_wrapper.ErrorWrapper
+	Update(branchID, itemID string, currentStock float64) *error_wrapper.ErrorWrapper
 	Delete(branchID, itemID string) *error_wrapper.ErrorWrapper
 }
